migrate governor and arbitrator hooks
