+++ src/util/get-appeal-cost.tsx
@@ -1,7 +1,9 @@
 import Web3 from "web3";
 
-export default (appealCost, multiplier, multiplierDivisor) => {
-  if (!appealCost || !multiplier || !multiplierDivisor) return "0";
+export const getAppealCost = (appealCost: number, multiplier: number, multiplierDivisor: number) => {
+  if (!appealCost || !multiplier || !multiplierDivisor) {
+    return "0";
+  }
 
   const toBN = Web3.utils.toBN;
